+++ package/models/model_grid_search.py
@@ -3,11 +3,11 @@ import json
 import logging
 import multiprocessing
 import os
-from typing import Dict, List, Tuple
-from pandas.core.indexes.numeric import Int64Index
+from typing import Dict, Tuple
 
 import pandas as pd
 from pandas.core.frame import DataFrame
+from pandas.core.indexes.numeric import Int64Index
 from sklearn.model_selection import GridSearchCV
 
 from data_split import DataSplit
@@ -57,6 +57,7 @@ class GridSearch(DataSplit):
             param_grid=self.gs_params,
             cv=self.train_test_index(),
             n_jobs=multiprocessing.cpu_count(),
+            scoring="neg_root_mean_squared_error",
             verbose=2,
         )
 
@@ -88,11 +89,8 @@ class GridSearch(DataSplit):
         elif self.model_type == ModelType.WITH_FEATURES:
             gs_path = os.path.join(self.model_path, "with_features")
 
-        self.gs_result.to_csv(os.path.join(gs_path, "gs_results.csv"))
+        self.gs_result.to_csv(os.path.join(gs_path, "gs_results.csv"), inedx=False)
 
         gs_best_params = json.dumps(self.gs.best_params_)
         with io.open(os.path.join(gs_path, "best_params.json"), "w") as f:
             f.write(gs_best_params)
-
-    def grid_search_splitter(self):
-        log.info("Start")